Fix mixEMT for failing samples and bugfixes
Fix mixEMT for failing samples by intercepting the exit code of mixEMT.
When mixEMT fails, check whether the log file is empty and the last line
of the log file is "0 contributors passed filtering steps.". If true,
replace the log file with a dummy one with NAs.

Bugfixes:
- fix ldpaths issue when installing a conda environment with snakemake
(relevant for contamMix)
- fix typos in README